Completed refactor of speed test logic into service
