+++ src/app/modules/speed-test/components/speed-statistics-preview/speed-statistics-preview.component.html
@@ -1,6 +1,6 @@
 <div class='statistics-container'>
-    <div>Words: {{ wordCount | number }}</div>
-    <div>Mistakes: {{ mistakeWordCount | number }}</div>
-    <div>WPM: {{ avgWpm | number:'1.1-1' }}</div>
-    <div>Accuracy: {{ avgAccuracy | number:'1.0-0' }}</div>
+    <div>Words: {{ typingStats.wordsTyped | number }}</div>
+    <div>Mistakes: {{ typingStats.incorrectWordCount | number }}</div>
+    <div>WPM: {{ typingStats.avgWpm | number:'1.1-1' }}</div>
+    <div>Accuracy: {{ typingStats.avgAccuracy | number:'1.0-0' }}</div>
 </div>
