+++ src/app/modules/speed-test/models/typing-stats.ts
@@ -1,0 +1,9 @@
+export class TypingStats {
+    wordsTyped: number = 0;
+    numLettersTyped: number = 0;
+    numIncorrectLetters: number = 0;
+    numExpectedLetters: number = 0;
+    incorrectWordCount: number = 0;
+    avgWpm: number = 0;
+    avgAccuracy: number = 0;
+}
