+++ src/app/modules/speed-test/services/speed-test.service.ts
@@ -3,6 +3,8 @@ import { WordLoaderService } from '../../core/services/word-loader.service';
 import { WordListApiService } from '../../core/services/word-list-api.service';
 import { TypingTestApiService } from '../../core/services/typing-test-api.service';
 import { TypingTestMode } from '../../core/models/typing-test-mode';
+import { TypingStats } from '../models/typing-stats';
+import { TypedWord } from '../models/typed-word';
 
 @Injectable()
 export class SpeedTestService {
@@ -10,6 +12,9 @@ export class SpeedTestService {
     selectedSpeedTestId: number = 1;
     words: string[] = [];
     lines: any[] = [];
+    currentWordIndex = 0;
+    typingStats: TypingStats = new TypingStats();
+    startTime: number = undefined;
 
     constructor(private _wordLoader: WordLoaderService,
                 private _wordListApiService: WordListApiService,
@@ -20,7 +25,40 @@ export class SpeedTestService {
     reset(): void {
         this.words = [];
         this.lines = [];
+        this.startTime = undefined;
+        this.typingStats = new TypingStats();
         this.loadWords();
+    }
+
+    wordTyped(word: TypedWord): void {
+        const incorrectLetters = this.calcIncorrectLetters(this.getCurrentWord(), word.word);
+        this.typingStats.numLettersTyped += word.word.length;
+        this.typingStats.numIncorrectLetters += incorrectLetters;
+        this.typingStats.numExpectedLetters += this.getCurrentWord().length;
+        if (incorrectLetters > 0) { this.typingStats.incorrectWordCount++; }
+        this.typingStats.wordsTyped++;
+
+        if (this.currentWordIndex === this.lines[0].length - 1) {
+            this.onLineCompleted();
+            this.currentWordIndex = 0;
+        } else {
+            this.currentWordIndex++;
+        }
+    }
+
+    letterTyped(): void {
+        if (!this.startTime) { this.startTime = Date.now(); }
+        this.updateStats();
+    }
+
+    getCurrentWord(): string {
+        return this.lines[0][this.currentWordIndex];
+    }
+
+    updateStats(): void {
+        const minElapsed = (Date.now() - this.startTime) / 60000;
+        this.typingStats.avgWpm = ((this.typingStats.numLettersTyped / 4.7) - this.typingStats.incorrectWordCount) / minElapsed;
+        this.typingStats.avgAccuracy = 100 - ((this.typingStats.numIncorrectLetters / this.typingStats.numExpectedLetters) * 100);
     }
 
     getNextWord(): string {
@@ -48,6 +86,17 @@ export class SpeedTestService {
     loadTest(testId: number): void {
         this.selectedSpeedTestId = testId;
         this.reset();
+    }
+
+    private calcIncorrectLetters(word: string, typedWord: string): number {
+        let incorrectCount = 0;
+        if (typedWord.length > word.length) { incorrectCount += (typedWord.length - word.length); }
+
+        for (let i = 0; i < Math.min(typedWord.length, word.length); i++) {
+            if (word.charAt(i) !== typedWord.charAt(i)) { incorrectCount++; }
+        }
+
+        return incorrectCount;
     }
 
     private loadTypingTests(): void {
